fix lit tag create
